backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
databricks-sql-connector>=3.1.0
redis>=5.0.0
python-dotenv>=1.0.0
